+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 * jre: published neowu/jre:26.0.1
 * stats: escalate to ERROR is cpu or memory high usage last more than 1 mins
   > call vm() and heap() every 1 min if memory usage is high, make it easier for AI troubleshoot
+* web: updated undertow to 2.4.1.Final
+* kafka: updated client to 4.3.0
 
 ### 9.5.2 (4/29/26 - 5/4/26)